Removed the ResponseMessage from the core module and made the ChatClientProviders generic, dependant on the individual implementation of the model provider
