+++ java/apirag-core/src/main/java/com/datasqrl/ai/models/ChatClientProvider.java
@@ -3,9 +3,9 @@ package com.datasqrl.ai.models;
 import java.util.List;
 import java.util.Map;
 
-public interface ChatClientProvider {
+public interface ChatClientProvider<Message> {
 
-  List<ResponseMessage> getChatHistory(Map<String, Object> context);
+  List<Message> getChatHistory(Map<String, Object> context);
 
-  ResponseMessage chat(String message, Map<String, Object> context);
+  Message chat(String message, Map<String, Object> context);
 }
