+++ java/apirag-openai/src/main/java/com/datasqrl/ai/models/openai/OpenAiChatProvider.java
@@ -4,20 +4,18 @@ import com.datasqrl.ai.backend.ChatSessionComponents;
 import com.datasqrl.ai.backend.FunctionBackend;
 import com.datasqrl.ai.backend.FunctionValidation;
 import com.datasqrl.ai.models.ChatClientProvider;
-import com.datasqrl.ai.models.ResponseMessage;
 import com.datasqrl.ai.util.JsonUtil;
 import com.theokanning.openai.completion.chat.*;
 import com.theokanning.openai.service.OpenAiService;
 
 import java.time.Duration;
-import java.time.Instant;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
-public class OpenAiChatProvider implements ChatClientProvider {
+public class OpenAiChatProvider implements ChatClientProvider<ChatMessage> {
 
   private final FunctionBackend backend;
   private final OpenAiService service;
@@ -33,7 +31,7 @@ public class OpenAiChatProvider implements ChatClientProvider {
   }
 
   @Override
-  public List<ResponseMessage> getChatHistory(Map<String, Object> context) {
+  public List<ChatMessage> getChatHistory(Map<String, Object> context) {
     OpenAIChatSession session = new OpenAIChatSession(model, systemPrompt, backend, context);
     List<ChatMessage> messages = session.retrieveMessageHistory(50);
     return messages.stream().filter(m -> {
@@ -42,11 +40,11 @@ public class OpenAiChatProvider implements ChatClientProvider {
         case USER, ASSISTANT -> true;
         default -> false;
       };
-    }).map(OpenAiChatProvider::toResponse).collect(Collectors.toUnmodifiableList());
+    }).collect(Collectors.toUnmodifiableList());
   }
 
   @Override
-  public ResponseMessage chat(String message, Map<String, Object> context) {
+  public ChatMessage chat(String message, Map<String, Object> context) {
     OpenAIChatSession session = new OpenAIChatSession(model, systemPrompt, backend, context);
     int numMsg = session.retrieveMessageHistory(20).size();
     System.out.printf("Retrieved %d messages\n", numMsg);
@@ -85,7 +83,7 @@ public class OpenAiChatProvider implements ChatClientProvider {
         FunctionValidation<ChatMessage> fctValid = session.validateFunctionCall(functionCall);
         if (fctValid.isValid()) {
           if (fctValid.isPassthrough()) { //return as is - evaluated on frontend
-            return toResponse(responseMessage);
+            return responseMessage;
           } else {
             System.out.println("Executing " + functionCall.getName() + " with arguments "
                 + functionCall.getArguments().toPrettyString());
@@ -96,7 +94,7 @@ public class OpenAiChatProvider implements ChatClientProvider {
         } //TODO: add retry in case of invalid function call
       } else {
         //The text answer
-        return toResponse(responseMessage);
+        return responseMessage;
       }
     }
   }
@@ -104,25 +102,5 @@ public class OpenAiChatProvider implements ChatClientProvider {
   public static Optional<ChatFunctionCall> getFunctionCallFromText(String text) {
     return JsonUtil.parseJson(text).map(json -> new ChatFunctionCall(json.get("function").asText(), json.get("parameters")));
   }
-
-  public static ResponseMessage toResponse(ChatMessage msg) {
-    ChatFunctionCall functionCall = null;
-    if (ChatMessageRole.valueOf(msg.getRole().toUpperCase()) == ChatMessageRole.ASSISTANT) {
-      functionCall = ((AssistantMessage) msg).getFunctionCall();
-    }
-    if (functionCall != null) {
-      System.out.println(functionCall.getArguments());
-      return new ResponseMessage(msg.getRole(), null,
-          functionCall.getArguments(), "", Instant.now().toString());
-    } else {
-      return new ResponseMessage(msg.getRole(),
-          msg.getTextContent(),
-          null,
-          "",
-          Instant.now().toString()
-      );
-    }
-  }
-
 
 }
