+++ java/apirag-server/src/main/java/com/datasqrl/ai/spring/ResponseMessage.java
@@ -1,4 +1,4 @@
-package com.datasqrl.ai.models;
+package com.datasqrl.ai.spring;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import lombok.AllArgsConstructor;
